requirements.txt
aiohttp==3.8.1
aiohttp-retry==2.5.2
aiosignal==1.2.0
alembic==1.7.3
altair==4.1.0
altair-data-server==0.4.1
altair-saver==0.5.0
altair-viewer==0.4.0
amqp==5.1.1
anyio==3.3.0
appdirs==1.4.4
argon2-cffi==20.1.0
async-generator==1.10
async-timeout==4.0.2
asyncssh==2.11.0
atpublic==3.0.1
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.10.0
billiard==4.0.0
bleach==4.1.0
blinker==1.4
bokeh==2.4.0
Bottleneck==1.3.2
brotlipy==0.7.0
cached-property==1.5.2
celery==5.3.0a1
certifi==2022.5.18.1
certipy==0.1.3
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudpickle==2.0.0
codetiming==1.3.0
colorama==0.4.4
commonmark==0.9.1
conda==4.10.3
conda-package-handling==1.7.3
configobj==5.0.6
copulas==0.7.0
cryptography==3.4.7
ctgan==0.5.1
cycler==0.10.0
Cython==0.29.24
cytoolz==0.11.0
dask==2021.9.1
debugpy==1.4.1
decorator==5.1.0
deepecho==0.3.0.post1
defusedxml==0.7.1
dictdiffer==0.9.0
dill==0.3.4
diskcache==5.4.0
distributed==2021.9.1
distro==1.7.0
docopt==0.6.2
dpath==2.0.6
dulwich==0.20.45
dvc==2.15.0
dvc-data==0.1.5
dvc-objects==0.1.5
dvc-render==0.0.6
dvc-task==0.1.0
dvclive==0.9.0
entrypoints==0.3
et-xmlfile==1.1.0
Faker==9.9.1
flatten-dict==0.4.2
flufl.lock==7.0
frozenlist==1.3.0
fsspec==2022.5.0
ftfy==6.1.1
funcy==1.17
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
gmpy2==2.1.0b5
grandalf==0.6
graphviz==0.20.1
greenlet==1.1.1
h11==0.13.0
h5py==3.4.0
HeapDict==1.0.1
idna==3.1
imagecodecs==2021.7.30
imageio==2.9.0
importlib-metadata==4.8.1
importlib-resources==5.2.2
iniconfig==1.1.1
ipykernel==6.4.1
ipympl==0.8.0
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==7.0.3
jupyter-core==4.8.1
jupyter-server==1.11.0
jupyter-telemetry==0.1.0
jupyterhub==1.4.2
jupyterlab==3.1.13
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
jupyterlab-widgets==1.0.2
kaleido==0.2.1
kiwisolver==1.3.2
kombu==5.3.0a1
llvmlite==0.37.0
locket==0.2.0
mailchecker==4.1.18
Mako==1.1.5
mamba==0.15.3
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
mock==4.0.3
mpmath==1.2.1
msgpack==1.0.2
multidict==6.0.2
nanotime==0.5.2
nbclassic==0.3.2
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5
notebook==6.4.4
numba==0.54.0
numexpr==2.7.3
numpy==1.20.3
oauthlib==3.1.1
olefile==0.46
openpyxl==3.0.10
outcome==1.2.0
packaging==21.0
pamela==1.0.0
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.2
partd==1.2.0
pathspec==0.9.0
patsy==0.5.2
pexpect==4.8.0
phonenumbers==8.12.51
pickleshare==0.7.5
Pillow==9.2.0
pip==21.2.4
pipreqs==0.4.11
plotly==5.9.0
pluggy==1.0.0
pooch==1.5.1
portpicker==1.5.2
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.18.0
psutil==5.9.1
psycopg2==2.8.6
ptyprocess==0.7.0
py==1.11.0
pycosat==0.6.3
pycparser==2.20
pycurl==7.44.1
pydot==1.4.2
pygit2==1.10.0
Pygments==2.10.0
pygtrie==2.5.0
PyJWT==2.1.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
pytest==7.1.2
python-benedict==0.24.3
python-dateutil==2.8.2
python-fsutil==0.6.1
python-json-logger==2.0.1
python-slugify==6.1.1
pyts==0.12.0
pytz==2022.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.3.0
rdt==0.6.4
requests==2.26.0
requests-unixsocket==0.2.0
rich==12.5.1
rpy2==3.4.5
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.2
ruamel-yaml-conda==0.15.80
scikit-image==0.18.3
scikit-learn==1.0
scipy==1.7.1
scmrepo==0.0.25
sdmetrics==0.5.0
sdv==0.16.0
seaborn==0.11.2
selenium==4.2.0
Send2Trash==1.8.0
setuptools==58.0.4
shortuuid==1.0.9
shtab==1.5.5
simplegeneric==0.8.1
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.0.1
SQLAlchemy==1.4.25
statsmodels==0.13.0
sympy==1.8
tables==3.6.1
tabulate==0.8.10
tblib==1.7.0
tenacity==8.0.1
terminado==0.12.1
testpath==0.5.0
text-unidecode==1.3
threadpoolctl==3.0.0
tifffile==2021.8.30
toml==0.10.2
tomli==2.0.1
toolz==0.11.1
torch==1.12.0
torchvision==0.13.0
tornado==6.1
tqdm==4.64.0
traitlets==5.1.0
trio==0.21.0
trio-websocket==0.9.2
typing-extensions==3.10.0.2
tzlocal==3.0
urllib3==1.26.7
vine==5.0.0
voluptuous==0.13.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
wheel==0.37.0
widgetsnbextension==3.5.1
wsproto==1.1.0
xlrd==2.0.1
xmltodict==0.13.0
yarg==0.1.9
yarl==1.7.2
zc.lockfile==2.0
zict==2.0.0
zipp==3.5.0
ipdb==0.13.9

openpyxl
inflection
PyYAML
imbalanced-learn
scikit-optimize
lightgbm
-e .